Enumerate, and route between, MIDI ports
A small CLI shell around two core features:

- MIDI port enumeration (`midi-tool list`)
- MIDI routing (`midi-tool route --source-name <output port> --target-name <input port>`)

Assuming you get the names of the ports right (do a `midi-tool list`
first), `midi-tool route` will start forwarding all MIDI events from one
port to the other. Play some notes on the source MIDI device and
hopefully the target device makes noises!

Non-functionally, there were a number of things I wanted to get out
of this first pass at the tool: getting acquainted with the language's
syntax; the semantics of borrowing and reference lifetimes; the standard
library and the Cargo ecossystem; and the feedback loop.

The CLI uses `clap`, which from what I could gather is a fairly standard
crate. Unsurprisingly so, since at least at the level of complexity we
have here the `#[derive(Parser)]` directive is a joy to use.

The MIDI stuff uses `midir`. In fact, the `route` subcommand is
basically a reimplementation of the `test_forward.rs` example bundled
with that crate. I worked up to parity using it as a reference while
navigating the foreignness of the language.

I plan on following up with additional MIDI features that I think of as
I use this when producing and jamming, maybe come up with a little swiss
knife type thing. Other than that, a few enhancements that I am
considering at the moment - keeping in mind this is a learning project:

- An interactive mode, with maybe a ncurses-based TUI (do people still
  use that these days?) where you can just navigate with the keyboard to
  create routings.

- Multithreading so the same process can forward multiple active routes
  in parallel.

- OS-native frontends for MacOS and Linux (Ghostty-style).