Validate args and set up routing with Devices
My original plan for this commit was to:

i. hide midir's MidiInput and MidiOutput objects behind Devices;

ii. have Devices provide a factory method which would build Route
objects, given a source/target port name pair.

iii. have Route be responsible for forwarding the MIDI messages, ie.
you'd call `activate` on a Route to do what we're doing with
`devices.activate(route)`.

I could do i. (see Devices::new), but not ii. nor iii..

Basically this factory method would take the `find_input_port` /
`find_output_port` logic from main, and it would output a Route if the
user parameters were valid. The `activate` method on the Route objects
would be called from the main module.

It turns out that struct Devices's ownership of the MidiInput /
MidiOutput objects complicates making Route responsible for forwarding.

Since I didn't have a particular reason for this, other than feeling
like this would be a reasonable OO design (the strict separation of
concerns promotes reusability when I get to the enhancements I have
planned), I decided it's fine to have Devices implement all of the logic
for now.

So I moved the route creation code in main to `Devices::activate`,
changed the signature so it takes the names as arguments, and renamed it
to `route` to accurately represent it's purpose.

To keep the design I had in mind, I think I'd have to make the Route
objects borrow the MidiInput and MidiOutput so they'd be able to call
`connect` on them. That's probably doable, maybe either:

- through the correct choice of lifetime identifiers for the references.
  Ie. Devices would have to outlive the Route instance so the references
  to the MidiIO objects don't disappear when Devices goes out of scope
  in `main`,

- or I'd implement the Copy trait for the MidiIO structs. I'm not sure
  if there's a reason it's not implemented out-of-the-box, it seems
  deliberate?