refactor: reuse routing to monitor input
`Devices::route()` and `Devices::monitor()` basically duplicated each
other - both read messages from the source port and process them with a
callback. This callback is where they differ. Whereas `route()` would
`send()` a message to an outgoing MIDI connection, `monitor()` would
print it to stdout.

I wanted to eliminate this duplication, making it possible to use a
Route struct and `Devices::route()` to implement monitoring/printing to
stdout.

My approach was to make MIDI message monitoring have the same API that
the midir crate provides for forwarding - you connect() to a MIDI device
and send() it MIDI messages, so you connect() to a monitor and send() it
messages to be printed. This entailed introducing a couple of wrapper
traits, Target (for connecting and describing for user feedback) and
Outgoing (sending messages).